Extract popup editing logic from content script
